Move SystemRuntime and StubActorRuntime to ailetos crate
This commit moves the core runtime infrastructure from the CLI binary
to the shared ailetos library crate, making it reusable across
different binaries.

Changes:
- Created ailetos/src/system_runtime.rs with SystemRuntime, IoRequest,
  Channel, FdTable, and related types
- Created ailetos/src/stub_actor_runtime.rs with StubActorRuntime
  (sync-to-async bridge for actors)
- Updated ailetos to re-export all runtime types for convenience
- Added tracing and futures dependencies to ailetos
- Removed ~755 lines of duplicate code from CLI main.rs
- Updated CLI to import runtime types from ailetos

All functionality remains identical with structured tracing in place.
