+++ ailets-rs/ailetos/Cargo.toml
@@ -16,6 +16,8 @@ tokio = { version = "1.41", features = ["sync", "time", "rt", "rt-multi-thread",
 log = "0.4"
 env_logger = "0.11"
 parking_lot = "0.12"
+tracing = "0.1"
+futures = "0.3"
 
 [[bin]]
 name = "pipe_demo"
